Refactor test to prepare multi-items flags.
