+++ include/orwell/com/RawMessage.hpp
@@ -21,5 +21,5 @@ public:
 
 };
 
-}} //namespace
+}}
